GET questions works
