+++ server/server.js
@@ -10,6 +10,7 @@ const passport = require('./strategies/user.strategy');
 // Route includes
 const userRouter = require('./routes/user.router');
 const testRouter = require('./routes/test.router');
+const questionRouter = require('./routes/question.router');
 const messageRouter = require('./routes/message.router');
 const eventRouter = require('./routes/event.router'); 
 const allUsersRouter = require('./routes/allusers.router');
@@ -28,6 +29,7 @@ app.use(passport.session());
 /* Routes */
 app.use('/api/user', userRouter);
 app.use('/api/test', testRouter);
+app.use('/api/question', questionRouter);
 app.use('/api/message', messageRouter);
 app.use('/api/event', eventRouter);
 app.use('/api/allusers', allUsersRouter)
