+++ src/redux/sagas/test.saga.js
@@ -13,8 +13,8 @@ function* testSaga() {
   yield takeLatest('UPDATE_TEST_SETTINGS', updateTestSettings);//updates any/all of the columns in the test table in the db.
     //dispatch({ type: 'UPDATE_TEST_SETTINGS', payload: { test: updatedTest } }); 
  
-  yield takeLatest('FETCH_QUESTIONS', fetchQuestions);//fetches all questions for the selected test.
-    // dispatch({ type: 'FETCH_QUESTIONS', payload: {test_id: test.id} });
+  yield takeLatest('FETCH_ALL_QUESTIONS', fetchAllQuestions);//fetches all questions for the selected test.
+    // dispatch({ type: 'FETCH_ALL_QUESTIONS', payload: {test_id: test.id} });
   yield takeLatest('ADD_QUESTION', addQuestion); //posts a new question to the db.
     //dispatch({ type: '', payload: { question: newQuestion} }
   yield takeLatest('UPDATE_QUESTION', updateQuestion); //updates an existing question in the db
@@ -30,21 +30,23 @@ function* deleteQuestion(action){
   //ap.question_id
   //ap.test_id
   try {
-    yield axios.delete(`/api/test/question/${ap.question_id}`);
+    yield axios.delete(`/api/question/${ap.question_id}`);
     yield put({ type: 'UNSET_SELECTED_QUESTION' });
-    yield put({ type: 'FETCH_QUESTIONS', payload: {test_id: ap.test_id}})
+    yield put({ type: 'FETCH_ALL_QUESTIONS', payload: {test_id: ap.test_id}})
 
   } catch (error) {
     console.log('DELETE question failed', error);
   }
 }
 
-// worker Saga: will be fired on "FETCH_QUESTIONS" actions
-function* fetchQuestions(action) {
+// worker Saga: will be fired on "FETCH_ALL_QUESTIONS" actions
+function* fetchAllQuestions(action) {
   const ap = action.payload;
   //ap.test_id 
   try {
-    const response = yield axios.get('/api/test/question', { params: {test_id: ap.test_id} } );
+    const response = yield axios.get('/api/question/all', { params: {test_id: ap.test_id} } );
+    console.log("response.data from fetch all qs:", response.data);
+
     yield put({ type: 'SET_ALL_QUESTIONS', payload: response.data });
   } catch (error) {
     console.log('get questions request failed', error);
@@ -58,7 +60,7 @@ function* updateQuestion(action){
   //id, point_value, type, required, question, option_one (thru six), answer, status, 
   //parent_test_id, created_by
   try {
-    yield axios.put(`/api/test/question/${ap.question.id}`, ap.question);
+    yield axios.put(`/api/question/${ap.question.id}`, ap.question);
     yield put({ type: 'SET-UPDATE_SELECTED_QUESTION', payload: ap.question });
       //todo ^ @Amanda - definitely need to verify that this works correctly
   } catch (error) {
@@ -74,9 +76,9 @@ function* addQuestion(action){
   //parent_test_id, created_by
   let question = ap.question;
   try {
-    const postedQuestion = yield axios.post('/api/test/question', ap.question );
+    const postedQuestion = yield axios.post('/api/question', ap.question );
     question = {...test, id: postedQuestion.data.id, create_date: postedQuestion.data.create_date, last_modified_date: postedQuestion.data.last_modified_date }
-    yield put({ type: 'FETCH_QUESTIONS', payload: {parent_test_id: parent_test_id} });
+    yield put({ type: 'FETCH_ALL_QUESTIONS', payload: {parent_test_id: parent_test_id} });
     yield put({ type: 'SET_SELECTED_QUESTION', payload: {question} })
   } catch (error) {
     console.log('POST question failed', error);
